refactor: remove use of CLI command on faux module
