+++ packages/salesforcedx-sobjects-faux-generator/src/describe/sObjectDescribe.ts
@@ -6,190 +6,10 @@
  */
 
 import { Connection } from '@salesforce/core';
-import {
-  CliCommandExecution,
-  CliCommandExecutor,
-  Command,
-  CommandOutput,
-  SfdxCommandBuilder
-} from '@salesforce/salesforcedx-utils-vscode/out/src/cli';
-import { extractJsonObject } from '@salesforce/salesforcedx-utils-vscode/out/src/helpers';
+import { DescribeGlobalResult, DescribeGlobalSObjectResult } from 'jsforce';
 import { CLIENT_ID } from '../constants';
-
-export interface SObject {
-  actionOverrides: any[];
-  activateable: boolean;
-  childRelationships: ChildRelationship[];
-  compactLayoutable: boolean;
-  createable: boolean;
-  custom: boolean;
-  customSetting: boolean;
-  deletable: boolean;
-  deprecatedAndHidden: boolean;
-  feedEnabled: boolean;
-  fields: Field[];
-  hasSubtypes: boolean;
-  isSubtype: boolean;
-  keyPrefix: string;
-  label: string;
-  labelPlural: string;
-  layoutable: boolean;
-  listviewable?: any;
-  lookupLayoutable?: any;
-  mergeable: boolean;
-  mruEnabled: boolean;
-  name: string;
-  namedLayoutInfos: any[];
-  networkScopeFieldName?: any;
-  queryable: boolean;
-  recordTypeInfos: RecordTypeInfo[];
-  replicateable: boolean;
-  retrieveable: boolean;
-  searchLayoutable: boolean;
-  searchable: boolean;
-  supportedScopes: SupportedScope[];
-  triggerable: boolean;
-  undeletable: boolean;
-  updateable: boolean;
-  urls: Urls2;
-}
-
-export interface ChildRelationship {
-  cascadeDelete: boolean;
-  childSObject: string;
-  deprecatedAndHidden: boolean;
-  field: string;
-  junctionIdListNames: any[];
-  junctionReferenceTo: any[];
-  relationshipName: string;
-  restrictedDelete: boolean;
-}
-
-export interface Field {
-  aggregatable: boolean;
-  autoNumber: boolean;
-  byteLength: number;
-  calculated: boolean;
-  calculatedFormula?: any;
-  cascadeDelete: boolean;
-  caseSensitive: boolean;
-  compoundFieldName?: any;
-  controllerName?: any;
-  createable: boolean;
-  custom: boolean;
-  defaultValue?: boolean;
-  defaultValueFormula?: any;
-  defaultedOnCreate: boolean;
-  dependentPicklist: boolean;
-  deprecatedAndHidden: boolean;
-  digits: number;
-  displayLocationInDecimal: boolean;
-  encrypted: boolean;
-  externalId: boolean;
-  extraTypeInfo?: any;
-  filterable: boolean;
-  filteredLookupInfo?: any;
-  groupable: boolean;
-  highScaleNumber: boolean;
-  htmlFormatted: boolean;
-  idLookup: boolean;
-  inlineHelpText?: any;
-  label: string;
-  length: number;
-  mask?: any;
-  maskType?: any;
-  name: string;
-  nameField: boolean;
-  namePointing: boolean;
-  nillable: boolean;
-  permissionable: boolean;
-  picklistValues: any[];
-  polymorphicForeignKey: boolean;
-  precision: number;
-  queryByDistance: boolean;
-  referenceTargetField?: any;
-  referenceTo: string[];
-  relationshipName: string;
-  relationshipOrder?: any;
-  restrictedDelete: boolean;
-  restrictedPicklist: boolean;
-  scale: number;
-  searchPrefilterable: boolean;
-  soapType: string;
-  sortable: boolean;
-  type: string;
-  unique: boolean;
-  updateable: boolean;
-  writeRequiresMasterRead: boolean;
-}
-
-export interface Urls {
-  layout: string;
-}
-
-export interface RecordTypeInfo {
-  active: boolean;
-  available: boolean;
-  defaultRecordTypeMapping: boolean;
-  master: boolean;
-  name: string;
-  recordTypeId: string;
-  urls: Urls;
-}
-
-export interface SupportedScope {
-  label: string;
-  name: string;
-}
-
-export interface Urls2 {
-  compactLayouts: string;
-  rowTemplate: string;
-  approvalLayouts: string;
-  uiDetailTemplate: string;
-  uiEditTemplate: string;
-  defaultValues: string;
-  describe: string;
-  uiNewRecord: string;
-  quickActions: string;
-  layouts: string;
-  sobject: string;
-}
-
-export interface DescribeSObjectResult {
-  result: SObject;
-}
-
-export enum SObjectCategory {
-  ALL = 'ALL',
-  STANDARD = 'STANDARD',
-  CUSTOM = 'CUSTOM'
-}
-
+import { BatchRequest, BatchResponse, SObject, SObjectCategory } from './types';
 export const MAX_BATCH_REQUEST_SIZE = 25;
-
-type SubRequest = { method: string; url: string };
-type BatchRequest = { batchRequests: SubRequest[] };
-type SubResponse = { statusCode: number; result: SObject };
-type BatchResponse = { hasErrors: boolean; results: SubResponse[] };
-
-export class ForceListSObjectSchemaExecutor {
-  public build(type: string): Command {
-    return new SfdxCommandBuilder()
-      .withArg('force:schema:sobject:list')
-      .withFlag('--sobjecttypecategory', type)
-      .withJson()
-      .build();
-  }
-
-  public execute(projectPath: string, type: string): CliCommandExecution {
-    const execution = new CliCommandExecutor(this.build(type), {
-      cwd: projectPath
-    }).execute();
-    return execution;
-  }
-}
-
 export class SObjectDescribe {
   private connection: Connection;
   private readonly servicesPath: string = 'services/data';
@@ -203,25 +23,28 @@ export class SObjectDescribe {
     this.connection = connection;
   }
 
-  public async describeGlobal(
-    projectPath: string,
-    type: SObjectCategory
-  ): Promise<string[]> {
-    const forceListSObjectSchemaExecutor = new ForceListSObjectSchemaExecutor();
-    const execution = forceListSObjectSchemaExecutor.execute(projectPath, type);
-    const cmdOutput = new CommandOutput();
-    let result: string;
-    try {
-      result = await cmdOutput.getCmdResult(execution);
-    } catch (e) {
-      return Promise.reject(e);
-    }
-    try {
-      const sobjects = extractJsonObject(result).result as string[];
-      return Promise.resolve(sobjects);
-    } catch (e) {
-      return Promise.reject(result);
-    }
+  /**
+   * Method that returns a list of SObjects based on running a describe global request
+   * More info at https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/resources_describeGlobal.htm
+   * @param type SObjectCategory
+   * @returns string[] containing the sobject names
+   */
+  public async describeGlobal(type: SObjectCategory): Promise<string[]> {
+    const requestedDescriptions: string[] = [];
+    const allDescriptions: DescribeGlobalResult = await this.connection.describeGlobal();
+
+    allDescriptions.sobjects.forEach((sobject: DescribeGlobalSObjectResult) => {
+      const isCustom = sobject.custom === true;
+      if (
+        type === SObjectCategory.ALL ||
+        (type === SObjectCategory.CUSTOM && isCustom) ||
+        (type === SObjectCategory.STANDARD && !isCustom)
+      ) {
+        requestedDescriptions.push(sobject.name);
+      }
+    });
+
+    return requestedDescriptions;
   }
 
   public buildSObjectDescribeURL(sObjectName: string): string {
